Add files via upload
A working LR classifier for the IMDB dataset, it internally creates a 1250 IMDB subset and processes it. The batch file is for testing parameters that the Grid Search doesn't cover. The program outputs a .csv file of performance results. It also outputs a log file of the best_params_ of the Grid Search and a Confusion Matrix image file.